requirements.txt
streamlit>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=5.0.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0